rag_app_python/requirements/requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements/requirements-dev.txt pyproject.toml
#
aiohttp==3.8.6
    # via openai
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   httpcore
    #   starlette
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
black==23.10.1
    # via brag (pyproject.toml)
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
charset-normalizer==3.3.2
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via
    #   black
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   loguru
    #   tqdm
coloredlogs==15.0.1
    # via onnxruntime
fastapi==0.104.1
    # via brag (pyproject.toml)
fastembed==0.1.1
    # via
    #   brag (pyproject.toml)
    #   qdrant-client
filelock==3.16.0
    # via huggingface-hub
flatbuffers==23.5.26
    # via onnxruntime
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.9.0
    # via huggingface-hub
greenlet==3.0.1
    # via sqlalchemy
grpcio==1.59.2
    # via
    #   grpcio-tools
    #   qdrant-client
grpcio-tools==1.59.2
    # via qdrant-client
gunicorn==21.2.0
    # via brag (pyproject.toml)
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==0.18.0
    # via httpx
httpx[http2]==0.25.0
    # via qdrant-client
huggingface-hub==0.24.6
    # via brag (pyproject.toml)
humanfriendly==10.0
    # via coloredlogs
hyperframe==6.0.1
    # via h2
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
loguru==0.7.2
    # via brag (pyproject.toml)
lxml==4.9.3
    # via python-docx
mpmath==1.3.0
    # via sympy
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
numpy==1.26.1
    # via
    #   onnx
    #   onnxruntime
    #   qdrant-client
onnx==1.15.0
    # via fastembed
onnxruntime==1.16.1
    # via fastembed
openai==0.28.1
    # via brag (pyproject.toml)
packaging==23.2
    # via
    #   black
    #   gunicorn
    #   huggingface-hub
    #   onnxruntime
pathspec==0.11.2
    # via black
pillow==10.4.0
    # via brag (pyproject.toml)
platformdirs==3.11.0
    # via black
portalocker==2.8.2
    # via qdrant-client
protobuf==4.25.0
    # via
    #   grpcio-tools
    #   onnx
    #   onnxruntime
psycopg2-binary==2.9.9
    # via brag (pyproject.toml)
pugsql==0.2.4
    # via brag (pyproject.toml)
pydantic==2.4.2
    # via
    #   fastapi
    #   pydantic-settings
    #   qdrant-client
pydantic-core==2.10.1
    # via pydantic
pydantic-settings==2.0.3
    # via brag (pyproject.toml)
pypdf2==3.0.1
    # via brag (pyproject.toml)
pyreadline3==3.4.1
    # via humanfriendly
python-docx==1.0.1
    # via brag (pyproject.toml)
python-dotenv==1.0.0
    # via pydantic-settings
python-multipart==0.0.6
    # via brag (pyproject.toml)
pywin32==306
    # via portalocker
pyyaml==6.0.2
    # via huggingface-hub
qdrant-client[fastembed]==1.6.4
    # via brag (pyproject.toml)
regex==2023.10.3
    # via tiktoken
requests==2.31.0
    # via
    #   fastembed
    #   huggingface-hub
    #   openai
    #   tiktoken
ruff==0.1.3
    # via brag (pyproject.toml)
sentry-sdk==1.34.0
    # via brag (pyproject.toml)
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==1.4.50
    # via pugsql
starlette==0.27.0
    # via fastapi
sympy==1.12
    # via onnxruntime
tiktoken==0.5.1
    # via brag (pyproject.toml)
tokenizers==0.13.3
    # via fastembed
tqdm==4.66.1
    # via
    #   fastembed
    #   huggingface-hub
    #   openai
typing-extensions==4.8.0
    # via
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   python-docx
urllib3==1.26.18
    # via
    #   qdrant-client
    #   requests
    #   sentry-sdk
uvicorn==0.23.2
    # via brag (pyproject.toml)
win32-setctime==1.1.0
    # via loguru
yarl==1.9.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
